libs/nisqa/requirements.txt
PyYAML==6.0.1
soundfile==0.12.1
# Стоит раскомментировать только, если выделять библиотеку в отдельные проект
#torch==2.4.1
#torchaudio==2.4.1
tqdm==4.66.1
